requirements.txt
azure-ai-agents
azure-identity
python-dotenv
bs4
requests